# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Instagram app",
        "version": "0.1"
    },
    "paths": {
...
        "/api/post": {
            "post": {
                "tags": [
                    "Create Post"
                ],
                "summary": "Create new post",
                "operationId": "0fab13e4c0a6523342b708d6808b762c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "png,jpeg",
                                        "example": "image.png"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is a description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful creation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Get Posts"
                ],
                "summary": "Get all posts",
                "operationId": "ef5a1d6f52f42f2bee93544d302e5e55",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/{post}": {
            "get": {
                "tags": [
                    "Get Post"
                ],
                "summary": "Get information for specific post",
                "operationId": "d541500fddeeefc729d1cd2f58f9dc38",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "image.jpg"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "This is a description"
                                        },
                                        "likes": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "time": {
                                            "type": "string",
                                            "example": "12:00 01 Jan 2022"
                                        },
                                        "comments": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        },
                                        "is_liked": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "users_liked": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Delete Post"
                ],
                "summary": "Delete post",
                "operationId": "c8ea09ed8f1261885a937483293bd900",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Id of post",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful creation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Comment": {
                "required": [
                    "id",
                    "content",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the comment",
                        "type": "integer",
                        "format": "int64"
                    },
                    "content": {
                        "description": "The content of the comment",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time when the comment was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the comment was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}